add update, delete and search products routes

+++ src/main/java/com/example/SpringBootFirstApp/controller/ProductController.java
@@ -1,14 +1,13 @@
 package com.example.SpringBootFirstApp.controller;
-
 import com.example.SpringBootFirstApp.model.Product;
 import com.example.SpringBootFirstApp.services.ProductService;
-import org.apache.coyote.Response;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
+import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
-
+import java.io.IOException;
 import java.util.List;
 
 @RestController
@@ -45,13 +44,49 @@ public class ProductController {
         }
     }
 
-    @PutMapping("/products")
-    public void updateProduct(@RequestBody Product product) {
-        service.updateProduct(product);
+    @DeleteMapping("/products/{id}")
+    public ResponseEntity<String> deleteProduct(@PathVariable int id) {
+        Product product = service.getProductById(id);
+        if(product == null) {
+            return new ResponseEntity<>("Product not found", HttpStatus.NOT_FOUND);
+        }
+        else{
+            service.deleteProduct(id);
+            return new ResponseEntity<>("Product deleted", HttpStatus.OK);
+        }
     }
 
-    @DeleteMapping("/products/{id}")
-    public void deleteProduct(@PathVariable int id) {
-        service.deleteProduct(id);
+    @GetMapping("/products/{productId}/image")
+    public ResponseEntity<byte[]> getProductImage(@PathVariable int productId) {
+        Product product = service.getProductById(productId);
+        if(product == null) {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
+
+        return ResponseEntity.ok()
+                .contentType(MediaType.valueOf(product.getImageType()))
+                .body(product.getImageData());
+
     }
+
+    @PutMapping("/products/{id}")
+    public ResponseEntity<String> updateProduct(@PathVariable int id, @RequestPart Product product, @RequestPart MultipartFile image) {
+        Product product1 = service.getProductById(id);
+        if(product1 == null) {
+            return new ResponseEntity<>( "Product not found",HttpStatus.NOT_FOUND);
+        }
+        try {
+            service.updateProduct(id, product, image);
+            return new ResponseEntity<>("Product Updated", HttpStatus.OK);
+        } catch (IOException e) {
+            return new ResponseEntity<>("Failed to update product", HttpStatus.INTERNAL_SERVER_ERROR);
+        }
+    }
+
+    @GetMapping("/products/search")
+    public ResponseEntity<List<Product>> searchProducts(@RequestParam("keyword") String keyword) {
+        List<Product> products = service.searchProducts(keyword);
+        return new ResponseEntity<>(products, HttpStatus.OK);
+    }
+
 }
